backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pandas>=2.1.3
pydantic>=2.5.2
pydantic-settings>=2.0.0
crewai>=0.1.0
python-multipart>=0.0.6
watchfiles>=0.21.0
langchain>=0.0.350
openai>=1.3.5
python-dotenv>=1.0.0